pyproject.toml
[project]
name = "github-history"
version = "0.1.0"
description = "Pipeline to fetch historical GitHub PR data and store in Supabase"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "pydantic>=2.12.4",
    "requests>=2.32.5",
    "supabase>=2.24.0",
]
